finish CRS-3

+++ src/main/java/pojo/customerService_2_0/CustomerService_2_0_InvalidMobilePhoneValue.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CustomerService_2_0_InvalidMobilePhoneValue {
-    private Integer mobilePhone;
+    private long mobilePhone;
 }
